requirements.txt
autopep8
bs4
requests
pyforms-gui
pyperclip